backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyMuPDF
pytesseract
pdf2image
chroma
openai
tqdm
streamlit
sentence-transformers
torch
requests
Pillow
chromadb
